Reorganize file strucucture
