+++ dklearn/__init__.py
@@ -1,1 +1,5 @@
+from __future__ import absolute_import
+
+from .core import from_sklearn
+
 __version__ = '0.1.0'
